test: add tests for postgres crud methods

+++ migrations/sqls/20221117165019-products-table-up.sql
@@ -2,6 +2,6 @@ CREATE TABLE products
 (
     id       SERIAL PRIMARY KEY,
     name     VARCHAR(128),
-    price    NUMERIC(11, 2),
+    price    INTEGER,
     category VARCHAR(32)
 );
